Reorganise & Update Controllers

+++ app/Http/Controllers/Api/AchievementController.php
@@ -7,44 +7,94 @@ use App\Http\Requests\Api\AchievementStoreRequest;
 use App\Models\Achievement;
 use Illuminate\Http\Request;
 
-
 class AchievementController extends Controller
 {
-    public function index(Request $request)
+    public function index()
     {
         $achievements = Achievement::all();
 
-        return response()->noContent(200);
+        if ($achievements->isEmpty()) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se Encontraron Logros!',
+                'data' => [],
+            ], 404);
+        }
+
+        return response()->json([
+            'status' => 'success',
+            'message' => '¡Logros Encontrados!',
+            'data' => $achievements,
+        ], 200);
     }
 
     public function store(AchievementStoreRequest $request)
     {
         $achievement = Achievement::create($request->validated());
 
-        return response()->noContent(201);
+        return response()->json([
+            'status' => 'success',
+            'message' => '¡Logro Creado Exitosamente!',
+            'data' => $achievement,
+        ], 201);
     }
 
-    public function show(Request $request, Achievement $achievement)
+    public function show(Achievement $achievement)
     {
-        return response()->noContent(200);
+        if (is_null($achievement)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Logro!',
+            ], 404);
+        }
+
+        return response()->json([
+            'status' => 'success',
+            'message' => '!Mostrando Datos del Logro!',
+            'data' => $achievement,
+        ], 200);
     }
 
     public function update(Request $request, Achievement $achievement)
     {
+        if (is_null($achievement)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Logro para Actualizar!',
+            ], 404);
+        }
+
         $achievement->update([]);
 
-        return response()->noContent(200);
+        return response()->json([
+            'status' => 'success',
+            'message' => '¡Logro Actualizado!',
+            'data' => $achievement,
+        ], 200);
     }
 
-    public function destroy(Request $request, Achievement $achievement)
+    public function destroy(Achievement $achievement)
     {
+        if (is_null($achievement)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Logro para Eliminar!',
+            ], 404);
+        }
+
         $achievement->delete();
 
-        return response()->noContent();
+        return response()->json([
+            'status' => 'success',
+            'message' => '¡Logro Eliminado!',
+        ], 204);
     }
 
-    public function error(Request $request)
+    public function error()
     {
-        return response()->noContent(400);
+        return response()->json([
+            'status' => 'error',
+            'message' => '¡Ha Ocurrido un Error con los Métodos del Controlador para Logros!',
+        ], 400);
     }
 }
