+++ app/Http/Controllers/Api/AddressController.php
@@ -5,45 +5,95 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\AddressStoreRequest;
 use App\Models\Address;
-use Illuminate\Http\Request;
 
 class AddressController extends Controller
 {
-    public function index(Request $request)
+    public function index()
     {
         $addresses = Address::all();
 
-        return response()->noContent(200);
+        if ($addresses->isEmpty()) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se Encontraron Direcciones!',
+                'data' => [],
+            ], 404);
+        }
+
+        return response()->json([
+            'status' => 'success',
+            'message' => '¡Direcciones Encontradas!',
+            'data' => $addresses,
+        ], 200);
     }
 
     public function store(AddressStoreRequest $request)
     {
         $address = Address::create($request->validated());
 
-        return response()->noContent(201);
+        return response()->json([
+            'status' => 'success',
+            'message' => '¡Dirección Creada Exitosamente!',
+            'data' => $address,
+        ], 201);
     }
 
-    public function show(Request $request, Address $address)
+    public function show(Address $address)
     {
-        return response()->noContent(200);
+        if (is_null($address)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado la Dirección!',
+            ], 404);
+        }
+
+        return response()->json([
+            'status' => 'success',
+            'message' => '!Mostrando Datos de la Dirección!',
+            'data' => $address,
+        ], 200);
     }
 
-    public function update(Request $request, Address $address)
+    public function update(Address $address)
     {
+        if (is_null($address)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado la Dirección para Actualizar!',
+            ], 404);
+        }
+
         $address->update([]);
 
-        return response()->noContent(200);
+        return response()->json([
+            'status' => 'success',
+            'message' => '¡Dirección Actualizada!',
+            'data' => $address,
+        ], 200);
     }
 
-    public function destroy(Request $request, Address $address)
+    public function destroy(Address $address)
     {
+        if (is_null($address)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado la Dirección para Eliminar!',
+            ], 404);
+        }
+
         $address->delete();
 
-        return response()->noContent();
+        return response()->json([
+            'status' => 'success',
+            'message' => '¡Dirección Eliminada!',
+        ], 204);
     }
 
-    public function error(Request $request)
+    public function error()
     {
-        return response()->noContent(400);
+        return response()->json([
+            'status' => 'error',
+            'message' => '¡Ha Ocurrido un Error con los Métodos del Controlador para Direcciones!',
+        ], 400);
     }
 }
